Refactorizando jsontransform a inyeccion de dependencias y dependencias de abstracciones
