+++ lib/components/json-transform/JsonTransform.type.ts
@@ -1,3 +1,5 @@
+import { ElementOpen } from "lib/shared/class/ElementOpen.csl";
+import { ITemplate } from "lib/shared/interfaces/Template.interface";
 import { ClassNames } from "lib/shared/types";
 
 export type JsonTransformType = ClassNames & {
@@ -11,4 +13,9 @@ export type JsonTransformType = ClassNames & {
   btnFormat: string;
   popup: string;
   errorInput: string;
+}
+
+export type JsonTransformServiceType = {
+  elementAttr: ElementOpen;
+  template: ITemplate<JsonTransformType>;
 }
